K: Add Student/Tutor account settings edit. Add log_in_as method for use in integration test (test/test_helper.rb). Add integration tests for Student/Tutor account settings edit. Streamline log in form. Remove subscribe button from Tutor's view of Course index (caused error).

+++ app/helpers/sessions_helper.rb
@@ -8,12 +8,12 @@ module SessionsHelper
   end
 
   def remember(user)
-    if user.class.name == Student.name
+    if student?
       user.remember_token = Student.new_token
       user.update_attribute(:remember_digest, Student.digest(user.remember_token))
       cookies.permanent.signed[:student_id] = user.id
       cookies.permanent[:remember_token] = user.remember_token
-    else
+    elsif tutor?
       user.remember_token = Tutor.new_token
       user.update_attribute(:remember_digest, Tutor.digest(user.remember_token))
       cookies.permanent.signed[:tutor_id] = user.id
@@ -32,18 +32,23 @@ module SessionsHelper
       @current_user ||= Student.find_by(id: user_id)
     elsif (user_id = session[:tutor_id])
       @current_user ||= Tutor.find_by(id: user_id)
+
     elsif (user_id = cookies.signed[:student_id])
+
       user = Student.find_by(id: user_id)
       if user && authenticated?(cookies[:remember_token], user)
         log_in_student user
         @current_user = user
       end
+
     elsif (user_id = cookies.signed[:tutor_id])
+
       user = Tutor.find_by(id: user_id)
       if user && authenticated?(cookies[:remember_token], user)
         log_in_tutor user
         @current_user = user
       end
+
     end
   end
 
@@ -58,20 +63,20 @@ module SessionsHelper
   # Forgets a persistent session.
   def forget(user)
     user.update_attribute(:remember_digest, nil)
-    cookies.delete(:student_id)
-    cookies.delete(:tutor_id)
+    cookies.delete(:student_id) if student?
+    cookies.delete(:tutor_id) if tutor?
     cookies.delete(:remember_token)
   end
 
-  # Logs out the current user.
   def logged_out?
     current_user.nil?
   end
 
+  # Logs out the current user.
   def log_out
     forget(current_user)
-    session.delete(:student_id)
-    session.delete(:tutor_id)
+    session.delete(:student_id) if student?
+    session.delete(:tutor_id) if tutor?
     @current_user = nil
   end
